Namespace root tracking working
This was a bad implementation. If namespaces aren't going to be in the heap, I need at least to track slots properly not loop over everything. I also don't love that namespaces are part of compiler instead of runtime. I probably need to find a better way to communicate to the runtime from the compiler.

+++ src/types.rs
@@ -273,7 +273,7 @@ impl HeapObject {
     pub fn get_field(&self, arg: usize) -> usize {
         let untagged = self.untagged();
         let pointer = untagged as *mut usize;
-        let pointer = unsafe { pointer.add(arg as usize + Self::header_size() / 8) };
+        let pointer = unsafe { pointer.add(arg + Self::header_size() / 8) };
         unsafe { *pointer }
     }
 }
